🤖 refactor: shelter type - api 폴더에서 types 폴더로 이동
