+++ src/api/shelter/admin/essential-info.ts
@@ -1,20 +1,5 @@
 import api from '@/api/instance';
-
-export type SearchedAddress = {
-  address: string;
-  postalCode: string;
-  latitude: number;
-  longitude: number;
-};
-export type ShelterAddress = SearchedAddress & {
-  addressDetail: string;
-};
-export interface ShelterEssentialInfo {
-  name: string;
-  phoneNumber: string;
-  description: string;
-  address: ShelterAddress;
-}
+import { ShelterEssentialInfo } from '@/types/shelter';
 
 export type ShelterEssentialInfoPayload = ShelterEssentialInfo;
 export type PutResponse = {
