+++ src/api/shelter/admin/index.ts
@@ -1,11 +1,5 @@
 import api from '@/api/instance';
-import { ShelterAdditionalInfo } from './additional-info';
-import { ShelterEssentialInfo } from './essential-info';
-
-export type ShelterInfo = ShelterEssentialInfo &
-  ShelterAdditionalInfo & {
-    profileImageUrl: string;
-  };
+import { ShelterInfo } from '@/types/shelter';
 
 export const get = async () => {
   return await api.get('shelter/admin').json<ShelterInfo>();
