+++ src/api/shelter/admin/useObservationAnimalList.ts
@@ -1,6 +1,7 @@
 import { UseQueryOptions, useQuery } from '@tanstack/react-query';
-import { ObservationAnimal, getAll } from './observation-animal';
+import { getAll } from './observation-animal';
 import { shelterKey } from '../../queryKey';
+import { ObservationAnimal } from '@/types/shelter';
 
 export default function useObservationAnimalList(
   options?: UseQueryOptions<ObservationAnimal[]>
