+++ src/api/shelter/admin/useShelterInfo.ts
@@ -1,6 +1,7 @@
 import { UseQueryOptions, useQuery } from '@tanstack/react-query';
 import { shelterKey } from '../../queryKey';
-import { ShelterInfo, get } from '.';
+import { get } from '.';
+import { ShelterInfo } from '@/types/shelter';
 
 export default function useShelterInfo(options?: UseQueryOptions<ShelterInfo>) {
   return useQuery<ShelterInfo>(shelterKey.info(), () => get(), {
